feat: add test

+++ dbAndCsvBatch/src/main/java/com/example/batch/enums/MemberType.java
@@ -6,12 +6,12 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum MemberType {
-  INDIVIDUAL(1, "Individual Member"),
-  BUSINESS(2, "Business Member"),
-  PREMIUM(3, "Premium Member"),
-  CORPORATE(4, "Corporate Member"),
-  GUEST(5, "Guest User");
+  INDIVIDUAL((byte) 1, "Individual Member"),
+  BUSINESS((byte) 2, "Business Member"),
+  PREMIUM((byte) 3, "Premium Member"),
+  CORPORATE((byte) 4, "Corporate Member"),
+  GUEST((byte) 5, "Guest User");
 
-  private final int value;
+  private final byte value;
   private final String description;
 }
